Backend/requirements.txt
fastapi
"uvicorn[standard]"
PyJWT
sqlalchemy
